Format

+++ tests/test_adjoint_jacobian.py
@@ -818,17 +818,17 @@ class TestAdjointJacobianQNode:
 
         assert np.allclose(grad_adjoint, grad_fd, atol=tol)
 
-@pytest.mark.parametrize("r_dtype,c_dtype", [[np.float32, np.complex64], [np.float64, np.complex128]])
+
+@pytest.mark.parametrize(
+    "r_dtype,c_dtype", [[np.float32, np.complex64], [np.float64, np.complex128]]
+)
 def test_qchem_expvalcost_correct(r_dtype, c_dtype):
     from pennylane import qchem
 
     symbols = ["Li", "H"]
     geometry = np.array([0.0, 0.0, 0.0, 0.0, 0.0, 2.969280527])
     H, qubits = qchem.molecular_hamiltonian(
-        symbols,
-        geometry,
-        active_electrons=2,
-        active_orbitals=5
+        symbols, geometry, active_electrons=2, active_orbitals=5
     )
     active_electrons = 2
     hf_state = qchem.hf_state(active_electrons, qubits)
@@ -840,10 +840,11 @@ def test_qchem_expvalcost_correct(r_dtype, c_dtype):
         qml.RZ(params[0], wires=2)
         qml.Hadamard(wires=1)
 
-    diff_method='adjoint'
-    custom_decomps={'DoubleExcitation': qml.DoubleExcitation.compute_decomposition}
-    dev_lig = qml.device("lightning.qubit", wires=qubits, c_dtype = c_dtype,
-            custom_decomps=custom_decomps)
+    diff_method = "adjoint"
+    custom_decomps = {"DoubleExcitation": qml.DoubleExcitation.compute_decomposition}
+    dev_lig = qml.device(
+        "lightning.qubit", wires=qubits, c_dtype=c_dtype, custom_decomps=custom_decomps
+    )
     cost_fn_lig = qml.ExpvalCost(circuit_1, H, dev_lig, optimize=False, diff_method=diff_method)
     circuit_gradient_lig = qml.grad(cost_fn_lig, argnum=0)
     params = np.array([0.123], requires_grad=True)
